Removing the upgrade fixes added in Version 2.1. These were determined to not be useful. While they were able to repair two issues we saw when upgrading Docker from 1.5 to 1.11, there was a third issue that just destroyed some data containers on update. So it was determined that the only safe way to update was to move all data out of containers before the upgrade and recreate all containers post-upgrade. With that in mind, it became clear that this code was adding complexity and no value here, so it was removed.
We also no longer pass the "--force-confnew" flag to apt when installing the Docker package.
This flag, added in 2.1.2, turned out to have some undesirable side effects - it meant that
any Docker config written before the Docker install would be erased during install, which
in turn meant we'd always have to start Docker with the wrong config, then immediately stop
it, fix the config, and start it again - not good.

This role now manages the "/etc/default/docker" file. This was done to allow us to set
Docker configuration that would be applied before startup.