Trust packages for whether reboot is necessary
Installing a kernel package on Ubuntu creates a
/var/run/reboot-required file. Key off its existence instead of trying
to figure out ourselves whether to reboot.

The `removes' argument to `command' handles conditioning the reboot on
whether the file exists.

For a vagrant instance, fire a `vagrant reload' after rebooting from
on the machine so we still get the setup `vagrant up' does.